fixed the imports and renamed test_admin_integration to mixins.py
