+++ tests/core/tests/admin_integration/mixins.py
