+++ tests/core/tests/admin_integration/test_views.py
@@ -1,3 +1,4 @@
+from core.tests.admin_integration.mixins import AdminTestMixin
 from django.contrib import admin
 from django.contrib.admin.options import ModelAdmin
 from django.contrib.auth.models import User
@@ -5,7 +6,6 @@ from django.test import RequestFactory
 from django.test.testcases import TestCase
 
 from import_export.admin import ImportExportMixinBase
-from tests.core.tests.admin_integration.test_admin_integration import AdminTestMixin
 
 
 class MockModelAdmin(ImportExportMixinBase, ModelAdmin):
